+++ Editor/Core/Data/ThunderKitSettings.cs
@@ -1,16 +1,10 @@
-﻿using System.Collections.Generic;
-using UnityEditor;
-using UnityEngine;
+﻿using UnityEditor;
 using UnityEditor.Compilation;
 using System.IO;
-using System;
 using System.Linq;
-using System.Reflection;
-using ThunderKit.Common;
 using ThunderKit.Core.Config;
 using ThunderKit.Markdown;
-using ThunderKit.Core.Windows;
-#if UNITY_2019 || UNITY_2020
+#if UNITY_2019_1_OR_NEWER
 using UnityEditor.UIElements;
 using UnityEngine.UIElements;
 #else